refector : 서비스 코드의 인터페이스 코드 분리
