+++ backend/src/main/java/com/isp/backend/domain/country/service/ExchangeRateScheduler.java
@@ -19,4 +19,5 @@ public class ExchangeRateScheduler {
             System.out.println("Failed to update exchange rates: " + e.getMessage());
         }
     }
+
 }
